client see each others messages
