+++ src/components/Login.js
@@ -1,26 +1,13 @@
 import React, { useRef } from 'react';
 import { Container, Form, Button } from 'react-bootstrap';
-// import { Socket as socket } from 'socket.io-client';
-import io from 'socket.io-client';
 
 function Login({ onUserNameSubmit }) {
-  const socket = io('http://localhost:3002');
-  socket.on('connect', () => console.log(`Hola ${socket.id}`));
-
   const userName = useRef();
 
   function handleSubmit(e) {
     e.preventDefault();
 
     onUserNameSubmit(userName.current.value);
-
-    const loggedUser = userName.current.value;
-    socket.auth = { userName };
-    // socket.auth = { loggedUser };
-    socket.connect();
-    socket.emit('Welcome Message', loggedUser);
-
-    socket.on('disconnect', () => 'server disconnected');
   }
 
   return (
